requirements.txt
starlette==0.37.2
uvicorn==0.29.0
tortoise-orm[aio-sqlite]==0.20.1
